+++ infrastructures/Kubernetes/Transcriber/CronJobTranscriber.php
@@ -159,7 +159,7 @@ class CronJobTranscriber implements DeploymentInterface
         bool $useHierarchicalNamespaces,
     ): TranscriberInterface {
         $requireLabel = $this->requireLabel;
-        $versionLevel = $this->versionLevel;
+        $versionLevel = $this->validateKubernetesVersionLevel($client);
         $sleepService = $this->sleepService;
 
         $compiledDeployment->foreachJob(