+++ infrastructures/Kubernetes/Transcriber/DeploymentTranscriber.php
@@ -153,7 +153,7 @@ class DeploymentTranscriber implements DeploymentInterface
         bool $useHierarchicalNamespaces,
     ): TranscriberInterface {
         $requireLabel = $this->requireLabel;
-        $versionLevel = $this->versionLevel;
+        $versionLevel = $this->validateKubernetesVersionLevel($client);
         $compiledDeployment->foreachPod(
             static function (
                 Pod $pod,