+++ infrastructures/Kubernetes/Transcriber/JobTranscriber.php
@@ -118,7 +118,7 @@ class JobTranscriber implements DeploymentInterface
         bool $useHierarchicalNamespaces,
     ): TranscriberInterface {
         $requireLabel = $this->requireLabel;
-        $versionLevel = $this->versionLevel;
+        $versionLevel = $this->validateKubernetesVersionLevel($client);
         $sleepService = $this->sleepService;
 
         $compiledDeployment->foreachJob(